feat(societal_values): Make societal value behave like estate satisfaction equilibrium
e.g. there is 75 Aristocracy & a trend towards Aristocracy of 0.5.

In vanilla, it stays at 75, until trend has different sign.
If it's less than that, it will increase to 50.
If it's more than 50, it maintains.

With this change, it will gravitate towards 50.
If a negative event brings it to 40, it will slowly go back to 50.