Make cost of the court generate market demand (#56)
For each country, the total money it spent on court expense will generate extra market demand in the market in which its capital is located.
The demand basket is modeled after the Royal Court building maintenance without Gold, Marble and Silver, since those could have limited supply. The actual demand size is scaled based on the current market price of the demand basket.

Closes #49

---------

Co-authored-by: Adelin <ademost95@yahoo.com>