chore: Add autosave file mover, to store autosaves (#118)
For future analysis this script can be placed in the savegames folder
and it will monitor the current directory for new Europa Universalis IV
autosave files (.eu5).

When a new autosave is detected, it reads the in-game date from the
file,renames the file to include the date,and moves it to a dedicated
"saves_watcher" directory.

It also manages the autosaves by retaining only those that are spaced at
least FREQUENCY_MONTHS apart, deleting closer ones to save space